refactor(slider): component optimization
