+++ packages/components/slider/__tests__/base.test.jsx
@@ -4,7 +4,7 @@ import { mount } from '@vue/test-utils';
 import { nextTick } from 'vue';
 import { describe, expect, it } from 'vitest';
 import Slider from '@src/slider/index.ts';
-import { formatLabel } from '@src/slider/util/common.ts';
+import { formatLabel } from '@src/slider/utils';
 
 // ui test
 describe('Slider', () => {
