+++ packages/components/slider/consts/index.ts
