+++ packages/components/slider/hooks/index.ts
@@ -1,0 +1,3 @@
+export * from './useSliderInput';
+export * from './useSliderMark';
+export * from './useSliderTooltip';
