+++ packages/components/slider/hooks/useSliderMark.tsx
@@ -3,7 +3,7 @@ import { cloneDeep } from 'lodash-es';
 import { SliderMarks } from '../type';
 import { TNode } from '../../common';
 import log from '@tdesign/common-js/log/log';
-import { getStopStyle } from '../util/common';
+import { getStopStyle } from '../utils';
 import TSliderMark from '../slider-mark';
 import { isArray } from 'lodash-es';
