+++ packages/components/slider/utils/index.ts
