+++ packages/components/time-picker/time-picker.tsx
@@ -24,7 +24,7 @@ dayjs.extend(customParseFormat);
 export default defineComponent({
   name: 'TTimePicker',
 
-  props: { ...props },
+  props,
 
   setup(props) {
     const renderTNodeJSX = useTNodeJSX();
